Add a provider for the attribute map
